// dist/allmanga.d.ts
type ChaptersForReadQueryExecutor = GQLExecutor<ChaptersForReadVariables, ChaptersConnection>;
type EpisodeVariables = OptionalExcept<EpisodeArguments, "showId" | "episodeString" | "translationType">;
type EpisodeQueryExecutor = GQLExecutor<EpisodeVariables, Episode>;
type EpisodeInfosVariables = OptionalExcept<EpisodeInfoArguments, "showId" | "episodeNumStart" | "episodeNumEnd">;
type EpisodeInfosQueryExecutor = GQLExecutor<EpisodeInfosVariables, EpisodeInfo[]>;
type MangaVariables = OptionalExcept<MangaArguments, "_id">;
type MangaQueryExecutor = GQLExecutor<MangaVariables, Manga>;
type MangasVariables = RequireOne<DeepPartial<MangasConnectionArguments>>;
type MangasQueryExecutor = GQLExecutor<MangasVariables, MangasConnection>;
type MangasWithdIdsVariables = OptionalExcept<MediaWithIdsArguments, "ids">;
type MangasWithdIdsQueryExecutor = GQLExecutor<MangasWithdIdsVariables, Manga[]>;
// ...
type ShowQueryExecutor = GQLExecutor<ShowVariables, Show>;
type ShowsVariables = Simplify<RequireOne<DeepPartial<ShowsConnectionArguments>>>;
type ShowsQueryExecutor = GQLExecutor<ShowsVariables, ShowsConnection>;
type ShowsWithIdsVariables = OptionalExcept<MediaWithIdsArguments, "ids">;
type ShowsWithIdsQueryExecutor = GQLExecutor<ShowsWithIdsVariables, Show[]>;
type AllMangaSDKUtils = {
	youtubeAnimeProxy: YoutubeAnimeProxy;
};
declare const utils: AllMangaSDKUtils;
type AllMangaSDK = {
	/** Schema Model `Manga`*/
	manga: MangaQueryExecutor;
	/** Schema Model `MangasConnection`*/
	mangas: MangasQueryExecutor;
	/** Schema Model `Manga`*/
	mangasWithIds: MangasWithdIdsQueryExecutor;
	/** Schema Model `Show`*/
	show: ShowQueryExecutor;
	/** Schema Model `ShowsConnection`*/
	shows: ShowsQueryExecutor;
	/** Schema Model `Show`*/
	showsWithIds: ShowsWithIdsQueryExecutor;
	/** Schema Model `ChaptersConnection`*/
	chaptersForRead: ChaptersForReadQueryExecutor;
	/** Schema Model `Episode`*/
	episode: EpisodeQueryExecutor;
	/** Schema Model `EpisodeInfo`*/
	episodeInfos: EpisodeInfosQueryExecutor;
	utils: typeof utils;
};
declare const allmanga: AllMangaSDK;